first-angular-app/src/app/page/login-page/login-page.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';

import { LoginPageComponent } from './login-page.component';

describe('LoginPageComponent', () => {
  let component: LoginPageComponent;
  let fixture: ComponentFixture<LoginPageComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [ LoginPageComponent ]
    })
    .compileComponents();

    fixture = TestBed.createComponent(LoginPageComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should verify username and password when login', function () {
    const spy = spyOn(component, 'verifyUser');
    component.user.username = 'none1';
    component.user.password = 'none2';
    component.submitForm();
    expect(spy).toHaveBeenCalledOnceWith('none1', 'none2');
  });
});
